refactor(glsl): move the emitter out of the translator
Writing the header is not a pass over the token list, and it sat among
the passes reading thirty odd fields of the translator and writing none
of them. It becomes Emitter, a record of exactly those answers, built at
render when all of them are settled: the header, the six prologues and
epilogues it calls, the output names, the opaque declarations and the two
that write an owed varying go with it. What the passes and the header
have to agree on, the names this engine injects, stays with the passes,
and the dozen constants and two records that spell them open to the
package for it.
